[#21090] Added the DebuggingDirectiveExtendedExampleSpec and refactored the code according to the comments
